[utils] Rename files to camelCase
