+++ src/core/account.ts
@@ -7,7 +7,7 @@ import { AccountAddress } from "./account_address";
 import { Hex } from "./hex";
 import { HexInput, SigningScheme } from "../types";
 import { PrivateKey, PublicKey, Signature } from "../crypto/asymmetric_crypto";
-import { derivePath } from "../utils/hd-key";
+import { derivePath } from "../utils/hdKey";
 import { AuthenticationKey } from "../crypto/authentication_key";
 import { Ed25519PrivateKey, Ed25519PublicKey } from "../crypto/ed25519";
 import { Secp256k1PrivateKey, Secp256k1PublicKey } from "../crypto/secp256k1";
