+++ tests/unit/transaction_builder.test.ts
@@ -26,7 +26,7 @@ import {
   signTransaction,
 } from "../../src/transactions/transaction_builder/transaction_builder";
 import { SigningScheme } from "../../src/types";
-import { Network } from "../../src/utils/api-endpoints";
+import { Network } from "../../src/utils/apiEndpoints";
 import { SignedTransaction } from "../../src/transactions/instances/signedTransaction";
 import { U64 } from "../../src/bcs/serializable/move-primitives";
 import { MoveObject } from "../../src/bcs/serializable/move-structs";
